requirements.txt
django == 3.0.10
django-pyodbc
django-pyodbc-azure
django == 3.0.10